DatabaseLoader Class

The `DatabaseLoader` class manages the database connection, loading configuration and metadata, and executing database operations in a specified order.

Details

This class provides methods to initialize a database connection, load the database with required data, and execute a series of operations defined in external R class scripts. The operations are executed in the order specified in the configuration file. The class also includes utility methods for cleaning up temporary files and retrieving all operations from a specified folder.

Methods

`initialize(db_path, config_path, cdm_metadata)`

Initializes the class with the provided parameters.

`db_path`: A string representing the path to the database file.

`config_path`: A string representing the path to configuration.

`cdm_metadata`: A string representing the path to the metadata.

`set_database()`

Loads the database with the required data.

`run_db_ops(ops)`

Executes a series of database operations in the order of config file.

`ops`: A list of operation objects to execute. If `NULL, they are retrieved using the `get_all_operations` method.

Public fields

db_path

A string representing the path to the database file.

data_instance

A string representing the path to the data instance.

config_path

A string representing the path to the configuration file.

cdm_metadata

A string representing the path to the metadata file.

config

A list containing the configuration from `config_path` file.

metadata

A `data.table` object containing the metadata loaded.

db

A database connection object created using the `duckdb` package.

Methods

Public methods

DatabaseLoader$new()

DatabaseLoader$set_database()

DatabaseLoader$run_db_ops()

DatabaseLoader$clone()

Method new()

Initializes the class with the provided parameters.

Usage

DatabaseLoader$new(
  db_path = NULL,
  data_instance = NULL,
  config_path = NULL,
  cdm_metadata = NULL
)

Arguments

db_path

A string representing the path to the database file.

data_instance

A string representing the path to the data instance.

config_path

A string representing the path to configuration file.

cdm_metadata

A string representing the path to the metadata file.

Method set_database()

Loads the database with the required data using config and metadata.

Usage

DatabaseLoader$set_database()

Method run_db_ops()

Executes a series of database operations in order specified in config.

Usage

DatabaseLoader$run_db_ops(ops = NULL)

Arguments

ops

A list of operation objects to execute. If `NULL`, operations are retrieved using the `get_all_operations` method.

Method clone()

The objects of this class are cloneable with this method.

Usage

DatabaseLoader$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
loader <- DatabaseLoader$new(
  db_path = "path/to/database.db",
  config_path = "path/to/set_db.json",
  cdm_metadata = "path/to/CDM_metadata.rds"
)
loader$set_database()
loader$run_db_ops()
} # }