Create Unique IDs for CDM Tables

This function generates a unique CDM ID for each data record of the specified Database CDM instance. This step is suggested to be applied after using the loadDb function. The function modifies the specified database by creating unique IDs for the specified CDM tables.

create_unique_id(
  db_connection,
  cdm_tables_names,
  extension_name = "",
  schema_name = NULL,
  to_view = FALSE,
  pipeline_extension = "_T2DMM"
)

Arguments

db_connection

Database connection object.

cdm_tables_names

List of CDM tables names to be imported into the db.

extension_name

String to be added to the name of the tables, useful when loading different CDM instances in the same database.

schema_name

Optional schema name to prepend to table and view names. Default is `NULL`.

to_view

Logical. If `TRUE` (default), creates a view with the unique ID column. If `FALSE`, overwrites the original table.

pipeline_extension

When using views when applying the clean_missing_values on CDM table we must define the name of the pipeline extension. See add_view for more information.

Examples

if (FALSE) { # \dontrun{
# Example usage of create_unique_id
db_connection <- dbConnect(duckdb::duckdb(), ":memory:")
cdm_tables_names <- c("PERSONS", "VISITS", "OBSERVATIONS")
create_unique_id(db_connection, cdm_tables_names,
  extension_name = "_CDM1"
)
} # }